libstagefright: Extract parts of FFMPEGSoftCodec for easier re-use
The code in Utils.cpp is also used in the MPEG2 extractor, which is now
compiled as a host library as well. As Utils.cpp was also using
FFMPEGSoftCodec, this introduced additional dependencies on the host
side (some of which were not available for host).

To avoid adding unnecessary dependencies on Utils.cpp, extract code from
FFMPEGSoftCodec into a separate file with less dependencies (same as
Utils.cpp), which can then be used in the extractor.